Player Health / Enemies
Made a little start on a basic concept enemy is just a circle atm :0. player will take damage every time it touches the enemy as it has a tag "enemies" so basically anything with that tag it will take damage too.  i need to add an end game function as currently player can get to 0 health and nothing happens :(.